Forget (for the moment) support for CF files.  Treat everything like NCAR. The problem with CF is that the interesting_names function rejects everything without a standard_name.  Transformed NCAR files say they follow the CF conventions and they do; but they don't use standard_name or any more than the bare minimum.
Soon we'll have to improve the whole system for identifying file types and, within
files, the variables worth indexing (probably most everything with an axis).
But presently all files we're testing on use NCAR formats or are from NCAR, so
this is the quick way to make it all work.