Add region functionality to AMWG plot set 5.
This requires adding region functionality to a couple of the
functions in reductions.py.

Note that region is implemented in the plot set as a class
attribute when the set is initialized, and can be called as
self.region.  Therefore it is not necessary to pass it to methods
in the function as an argument.