Repaired a problem with getSections in reductions, used in plot set 12.
Upgraded diags-new to be current and provide the same functionality as diags.
It successfully passes the tests that diags does.  The only ting left to do
before it becomes the default is to upgrade the the tests since it changes
the file names.

Made chaneges to amwg mostly in plot set 12.