Rename the amwg_plot_plan class variable standard_variables to common_derived_variables.  Move it to its own file, atmos_derived_vars.py.  Change names of related methods, etc. to agree with this new name.
In atmos_derived_vars there is a new class method get_user_vars (belongs to amwg_plot_plan).
The user can now define his own variables, just as common_derived_variables is defined, in his
own file.  This method will read the file and add the user variables to common_derived_variables.
This method is not called yet, but when it is called and documented, we will have a "tight coupling"
system for user-defined diagnostics, which does not require changes to the uvcmetrics code, and
in principle requires no involvement by uvcmetrics developers.